mocked the file exists the all tests passing

+++ lib/bitmap_editor.rb
@@ -7,7 +7,7 @@ class BitmapEditor
   end
 
   def run(file)
-    # return puts "please provide correct file" if file.nil? || !File.exists?(file)
+    return puts "please provide correct file" if file.nil? || !File.exists?(file)
 
     File.open(file).each do |line|
       line = line.chomp.split
